+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Fixed
 - Improve the disk space check error message to accurately reflect the space requirement when accounting for stored videos.
+- Add warning when delete selection contains locked videos.
 
 ## [7.7.2] - 2026-05-24
 ### Fixed